SmallLinting

+++ Algos/AStar/A_star.cpp
@@ -6,11 +6,9 @@
 
 
 bool operator < (const DistTo& lhs, const DistTo& rhs) {
-
 	return lhs.Dist_ < rhs.Dist_;
 }
 bool operator > (const DistTo& lhs, const DistTo& rhs) {
-
 	return lhs.Dist_ > rhs.Dist_;
 }
 
@@ -23,15 +21,15 @@ void Graph::Parse(std::istream& source)
 	int n, m;
 	source >> n >> m;
 	adj.resize(n);
-	for (int i = 0; i < n; i++)
+	for (size_t i = 0; i < n; i++)
 	{
 		int x, y;
 		source >> adj[i].coord.x >> adj[i].coord.y;		
 	}
-	for (int i = 0; i < m; i++) {
+	for (size_t i = 0; i < m; i++) {
 		int x, y, w;
 		source >> x >> y >> w;
-		adj[x - 1].AdjList[ y - 1] =w ;
+		adj[x - 1u].AdjList[ y - 1u] =w ;
 		
 	}
 }
@@ -40,7 +38,7 @@ void Graph::Process(std::istream& source, std::ostream& output)
 {
 	int t;
 	source >> t;
-	for (int i = 0; i < t; ++i) {
+	for (size_t i = 0; i < t; ++i) {
 		int u, v;
 		source >> u >> v;
 		output << AStar(u - 1, v - 1) << '\n';
@@ -59,14 +57,12 @@ long long Graph::AStar(int s, int t){
 	changed.push_back(s);
 	std::priority_queue <DistTo, std::vector <DistTo>, std::greater<DistTo>> H;
 	H.push({ s,0 });
-	while (!H.empty())
-	{
+	while (!H.empty()){
 		auto u = H.top();
 		H.pop();
 		DistTo current{ u.vertex_,(double)adj[u.vertex_].d };
 		if (u < current) continue;
-		for (auto v : adj[u.vertex_].AdjList)
-		{
+		for (auto v : adj[u.vertex_].AdjList){
 
 			if (adj[v.first].d > adj[u.vertex_].d + v.second)
 			{
@@ -81,8 +77,7 @@ long long Graph::AStar(int s, int t){
 	}
 	auto current = t;
 	long long dist = 0;
-	while (current != s)
-	{
+	while (current != s){
 		auto pi = adj[current].prev;
 		if (pi < 0) {
 			dist = -1;
@@ -101,8 +96,7 @@ long long Graph::AStar(int s, int t){
 
 
 
-void Graph::Vipe(const std::vector<int>& changed)
-{
+void Graph::Vipe(const std::vector<int>& changed){
 	for (auto item : changed)
 	{
 		adj[item].d = Infinity;
